4-isalpha

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -10,7 +10,7 @@ int main(void)
 
 {
 	int i = 0;
-	long j = i, k = 2, sum = k;
+	int j = i, k = 2, sum = k;
 
 	while (k + j < 400000)
 	{
